% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milags_functions.R
\name{choose_lag_fit_algorithm_baranyi}
\alias{choose_lag_fit_algorithm_baranyi}
\title{choose_lag_fit_algorithm_baranyi}
\usage{
choose_lag_fit_algorithm_baranyi(
  gr_curve,
  LOG10N0,
  init_lag,
  init_mumax,
  init_LOG10Nmax,
  max_iter,
  lower_bound
)
}
\arguments{
\item{gr_curve}{data from one specific growth curve with the following columns: LOG10N, t}

\item{LOG10N0}{init value for the LOG10N0 parameter}

\item{init_lag}{initial value for the lag}

\item{init_mumax}{initial value for the mumax parameter}

\item{init_LOG10Nmax}{initial value for the LOG10Nmax parameter}

\item{max_iter}{max. number of iterations}

\item{lower_bound}{lower bound for the bounded nls optimization;}
}
\value{
the best nls fitting object with parameters fitted to Baranyi model (lowest Res.Sum Sq provided that all coefficients are nonnegative)
}
\description{
Runs nlsLM/nls algorithms with three different parameter setups to fit the best Baranyi parameters to our data and chooses the best model
}
