% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milags_functions.R
\name{compare_algorithms}
\alias{compare_algorithms}
\title{compare_algorithms}
\usage{
compare_algorithms(nls_LM_no_bound, nls_PORT, nlsres_LM)
}
\arguments{
\item{nls_LM_no_bound}{first object resulting from running nls}

\item{nls_PORT}{second object resulting from running nls}

\item{nlsres_LM}{third object resulting from running nls}
}
\value{
the best fitting object (lowest Res.Sum Sq provided that all coefficients are nonnegative)
}
\description{
Compares results of 3 objects obtained from running nls
}
