% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milags_functions.R
\name{cut_the_data}
\alias{cut_the_data}
\title{cut_the_data
Subsets the data frame containing only the observations up to the specified maximum time}
\usage{
cut_the_data(data, max_time)
}
\arguments{
\item{data}{a data frame with two required columns names: "time" and "biomass",and one optional column: "curve_id"
This is data from may come from multiple growth curves}

\item{max_time}{max. time at which we want to cut the growth curve data}
}
\value{
cut data
}
\description{
cut_the_data
Subsets the data frame containing only the observations up to the specified maximum time
}
