% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milags_functions.R
\name{fit_max_infl_lag}
\alias{fit_max_infl_lag}
\title{fit_max_infl_lag}
\usage{
fit_max_infl_lag(data)
}
\arguments{
\item{data}{a data frame with two required columns names: "time" and "biomass",and one optional column: "curve_id"
This is data from may come from multiple growth curves}
}
\value{
growth curve data (as input) together with additional columns: lag, log.biomass, time.diff, time.av, second.deriv.b, biomass.increase
}
\description{
Fits the lag to multiple growth curves based on the max growth acceleration method
It finds where the second derivative is the largest
}
