% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milags_functions.R
\name{get_def_pars}
\alias{get_def_pars}
\title{get_def_pars
Set defaults parameters used by calc_lag function}
\usage{
get_def_pars()
}
\value{
list of parameters
}
\description{
get_def_pars
Set defaults parameters used by calc_lag function
}
