% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milags_functions.R
\name{get_n0}
\alias{get_n0}
\title{get_n0}
\usage{
get_n0(biomass, n0_method)
}
\arguments{
\item{biomass}{vector of biomass (chronologically ordered as in growth curve)}

\item{n0_method}{"first.observation" if the first point is taken as the initial biomass or
"minimal.observation" if the minimal biomass is taken is the initial point.
In "healthy" growth curves these options should be equivalent
but sometimes a drop in OD/biomass is observed at the beginning of a growth curve.
In this case it is not obvious what to assume the initial biomass is.}
}
\value{
a value of the initial biomass (either the first observation or the minimum value depending on the parameter N0.method)
}
\description{
Gets the initial biomass to relate to
}
