% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milags_functions.R
\name{get_theme}
\alias{get_theme}
\title{get_theme}
\usage{
get_theme(text_size = 12)
}
\arguments{
\item{text_size}{defaults to 12}
}
\value{
a ggplot theme
}
\description{
This function sets a ggplot theme without grid.
The theme removes the major and minor grid lines, sets a white background with a gray border and adjusts the text size.
}
