% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milags_functions.R
\name{lag_biomass_incr}
\alias{lag_biomass_incr}
\title{lag_biomass_incr}
\usage{
lag_biomass_incr(data, threshold, n0)
}
\arguments{
\item{data}{a data frame with two required columns names: "time" and "biomass",and one optional column: "curve_id"
This is data from may come from multiple growth curves}

\item{threshold}{A value of the biomass increase that we can surely associate with the end of the lag phase rather than random variation during the lag}

\item{n0}{the initial biomass (lag will be defined as the time point where the difference between biomass and N0 reaches a predefined threshold)}
}
\value{
growth curve data (as input) together with additional columns: N0, increase.from.N0, lag
}
\description{
Fits the lag to multiple growth curves based on the biomass increase method
}
