% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milags_functions.R
\name{make_grwoth_curve_df}
\alias{make_grwoth_curve_df}
\title{make_grwoth_curve_df}
\usage{
make_grwoth_curve_df(time, biomass, curve_id = NULL)
}
\arguments{
\item{time}{numeric vector of times when biomass was measured  (chronologically ordered as in growth curve)}

\item{biomass}{numeric vector of measured biomass values  (chronologically ordered as in growth curve)}

\item{curve_id}{character vector of growth curve identifiers (i.e. if there are multiple measurements done at the same time point, they should have different curve_id)}
}
\value{
a data frame representing growth curve data
}
\description{
Create a growth curve data frame that can be later passed to the lag clculation functions
}
