% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milags_functions.R
\name{smooth_data}
\alias{smooth_data}
\title{smooth_data
Smoothens growth curves data}
\usage{
smooth_data(data, smooth_kind = "3RS3R")
}
\arguments{
\item{data}{a data frame with two required columns names: "time" and "biomass",and one optional column: "curve_id"
This is data from may come from multiple growth curves}

\item{smooth_kind}{kind used for the smooth functions, defaults to "3RS3R"}
}
\value{
smoothened data
}
\description{
smooth_data
Smoothens growth curves data
}
